backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pymongo==4.9.2
joblib==1.4.2
python-dotenv==1.0.1
